// whatsapp-backend/src/core/fsm/enums/BotState.ts
/**
 * @file BotState.ts
 * @description Enumeración de los estados finitos y deterministas de la conversión.
 */
export enum BotState {
  WELCOME = 'WELCOME',
  AWAITING_NAME_CAPTURE = 'AWAITING_NAME_CAPTURE',
  AWAITING_DOCUMENT_CAPTURE = 'AWAITING_DOCUMENT_CAPTURE',
  AWAITING_MENU_OPTION = 'AWAITING_MENU_OPTION',
  AWAITING_MENU_SELECTION = 'AWAITING_MENU_SELECTION',
  AWAITING_ORIGIN_CAPTURE = 'AWAITING_ORIGIN_CAPTURE',
  AWAITING_DESTINATION_CAPTURE = 'AWAITING_DESTINATION_CAPTURE',
  AWAITING_WEIGHT_CAPTURE = 'AWAITING_WEIGHT_CAPTURE',
  AWAITING_VEHICLE_TYPE_SELECTION = 'AWAITING_VEHICLE_TYPE_SELECTION',
  HUMAN_INTERCEPTION = 'HUMAN_INTERCEPTION',
  ASKING_ORIGIN = 'ASKING_ORIGIN',
  ASKING_DESTINATION = 'ASKING_DESTINATION',
  COMPUTING_LIQUIDATION = 'COMPUTING_LIQUIDATION'
}